requirements.txt
colorama==0.4.6
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
psutil==7.0.0
pyperclip==1.9.0
pywin32==308
qrcode==8.0
ttkbootstrap==1.10.1
